Initial commit
Like the readme says, I'm looking for a space opera rpg, because I want
to build my own settings for it. This is the start of that project.

What have we now? Just the starship encounter mechanics. Look, I want this
thing to be lightweight, because otherwise I'll get annoyed at all the damn
rules I've written, which is why the starship mechanics fits in 81 lines at
78 columns. Not all that much text.

If you want FTL stuff, make it yourself. The setting (which I'll eventually
get to) will have it's own explaination for why starships don't have their
own FTL capabilities.