The License. Oh the License...
I picked the smile license, because I want people to use this stuff as
they will, but also be nice to somebody.

Basically don't be an ass. Unless you're the GM, in which case, be the
biggest ass you can possibly be to your players, because it makes life
interesting.

Did I mention, one time, in a Star Wars campaign I ran, I had one player
who pretty much would do anything for their home world, and decided that
said home world needed to be next in line for the Yuuzhan Vong invasion?
Yeah, that was a fun little side adventure. That player stole a ship from
the rest of the party, than decided to split the party. Good thing the
other two players had a way to disable that ship. They than neatly blew
the traitor out the airlock.

I wonder if anybody else ever reads these things, or if they're just for
me. Dear anybody who reads the entire commit message on this project:
congratulations, you've won nothing but some slightly wasted time and a
fun little easter egg of sorts. To collect your prize, continue reading.

Oh wait, it's over...