Slingshots (the FTL tech) and a New Hope explained
The Slingshot technology, this universe's method of FTL transportation
is now explained, along with the first planet, New Hope (yes, named for
the Star Wars movie, at least, that's the out-of-universe explaination,
the in-universe explaination probably involves some heart-breaking story
of the first settlers attempts to find some sort of new hope on this
planet. Yawn).

Basically, the thing that lets people get from place to place quickly,
along with an okay starting point for an adventure or campaign.

Look, I don't know why anybody else is reading this, but I know why I'm
writing it. I want a generic, rules-lite, space opera, table-top
roleplaying game. I'm just writing the setting stuffs for my own uses.

For people looking for some insight as to how I'm going to be writing
things, look no further than the rest of this commit message!

Basically, I've been writing out broad ideas on a notepad. I've also
been writing some adventure hooks, and in some cases, specific mechanics
for how things work in this universe. After getting my initial thoughts
down, I've come to this stuff, and started, first expanding on my
mechanics, followed by the setting that might (or might not) use those
mechanics, followed by the adventure that uses the setting and mechanics.
What this means right now is that up next is either another planet
(this time Quorat) or the adventure that utilizes the two pieces of setting
and the one mechanic I've already written. It's called Abbandoned. It'll be
FUN. Because I said so.